fetch/metadata/generated/cors-preflight-sec-fetch-dest.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Dest</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['crossSite'], { requireHEAD: true });

    return fetch(url, { method: 'DELETE' })
      // This request is expected to fail
      .catch(() => {})
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'empty');
        });
  }, 'sec-fetch-dest - ');
  </script>
  </body>
</html>
